First attempt at implementing the fibonacci
This works, but is not performant - bump up _n_ to 200 and it'll grind
your CPU. I left it running for about 30 - 60 mins and it didn't
complete.

I committed this mainly to show the thinking I went down first. I will
say that the "hint" for this one threw me a bit since it suggested that
_...you might need two values..._. So I was looking for a solution that
only needed two params.