ex3.6 - First attempt without looking at answer.
I figured that you need to keep track of the items in the list that
you've processed. To do that I guess you need a little loop method and a
new variable. We're basically checking to see if the next item in the
list is `Nil`. We start with an empty list and if the next item in the
list is not `Nil` then we add the `h` to our current list and then pass
on the next segment in the list to the recursive loop. If the list item
we're checking's tail is `Nil` then we just return what we've build so
far.

This function can't be executed in constant time because you need to
traverse the entire list to work it out. So the time for the function to
execute is totally bound to the length of the list.